thermal: msm_lmh_dcvs: Never initialize LMH TZ more than once
The LMH DCVS initialization consists in enabling the LMH thermal,
outer loop, reliability and BCL algorithms on a per-cluster
basis, but the command that kickstarts it is "PROFILE_CHANGE"
with parameter '1' sent to the TZ LMH Service.

Sending it more than one time is rendundant, but the critical
issue is that - if we send it more than once - the CPUs will
deadlock!

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Ie077fe1562ba67393db03d53bf5dc2ff8efc6fb8
Signed-off-by: TogoFire <togofire@mailfence.com>
Signed-off-by: Sorayukii <sorayukii69@gmail.com>
Signed-off-by: Abdallah Ibrahim <iabdollah40@gmail.com>